+++ Sources/CodingAgentKit/CodingAgentKit.swift
@@ -1,3 +1,4 @@
 @_exported import AgentCore
 @_exported import ClaudeCodeKit
+@_exported import DelegateKit
 @_exported import OpenCodeKit